+++ packages/jupyter-ai/package.json
@@ -62,10 +62,8 @@
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
     "@jupyter/chat": "^0.6.0",
-    "@jupyter/collaboration": "^1",
     "@jupyterlab/application": "^4.2.0",
     "@jupyterlab/apputils": "^4.2.0",
-    "@jupyterlab/cells": "^4.2.0",
     "@jupyterlab/codeeditor": "^4.2.0",
     "@jupyterlab/codemirror": "^4.2.0",
     "@jupyterlab/completer": "^4.2.0",
