+++ packages/jupyter-ai/src/tokens.ts
@@ -1,12 +1,9 @@
 import React from 'react';
 import { Token } from '@lumino/coreutils';
 import { ISignal } from '@lumino/signaling';
-import type { IRankedMenu, ReactWidget } from '@jupyterlab/ui-components';
+import type { IRankedMenu } from '@jupyterlab/ui-components';
 
 import { AiService } from './handler';
-import { ChatHandler } from './chat_handler';
-import { ActiveCellManager } from './contexts/active-cell-context';
-import { SelectionWatcher } from './selection-watcher';
 
 export interface IJaiStatusItem {
   addItem(item: IRankedMenu.IItemOptions): void;
@@ -49,23 +46,6 @@ export interface IJaiMessageFooter {
 export const IJaiMessageFooter = new Token<IJaiMessageFooter>(
   'jupyter_ai:IJaiMessageFooter',
   'Optional component that is used to render a footer on each Jupyter AI chat message, when provided.'
-);
-
-export interface IJaiCore {
-  chatWidget: ReactWidget;
-  chatHandler: ChatHandler;
-  activeCellManager: ActiveCellManager;
-  selectionWatcher: SelectionWatcher;
-}
-
-/**
- * The Jupyter AI core provider token. Frontend plugins that want to extend the
- * Jupyter AI frontend by adding features which send messages or observe the
- * current text selection & active cell should require this plugin.
- */
-export const IJaiCore = new Token<IJaiCore>(
-  'jupyter_ai:core',
-  'The core implementation of the frontend.'
 );
 
 /**
