+++ packages/jupyter-ai/src/utils.ts
@@ -1,39 +1,34 @@
 /**
  * Contains various utility functions shared throughout the project.
  */
-import { Notebook } from '@jupyterlab/notebook';
-import { FileEditor } from '@jupyterlab/fileeditor';
 import { CodeEditor } from '@jupyterlab/codeeditor';
+import { CodeMirrorEditor } from '@jupyterlab/codemirror';
+import { DocumentWidget } from '@jupyterlab/docregistry';
+import { FileEditor } from '@jupyterlab/fileeditor';
+import { Notebook } from '@jupyterlab/notebook';
 import { Widget } from '@lumino/widgets';
 
 /**
- * Get text selection from an editor widget (DocumentWidget#content).
- */
-export function getTextSelection(widget: Widget): string {
-  const editor = getEditor(widget);
-  if (!editor) {
-    return '';
-  }
-
-  const selectionObj = editor.getSelection();
-  const start = editor.getOffsetAt(selectionObj.start);
-  const end = editor.getOffsetAt(selectionObj.end);
-  const text = editor.model.sharedModel.getSource().substring(start, end);
-
-  return text;
-}
-
-/**
- * Get editor instance from an editor widget (i.e. `DocumentWidget#content`).
+ * Gets the editor instance used by a document widget. Returns `null` if unable.
  */
 export function getEditor(
-  widget: Widget
-): CodeEditor.IEditor | null | undefined {
+  widget: Widget | null
+): CodeMirrorEditor | null | undefined {
+  if (!(widget instanceof DocumentWidget)) {
+    return null;
+  }
+
   let editor: CodeEditor.IEditor | null | undefined;
-  if (widget instanceof FileEditor) {
-    editor = widget.editor;
-  } else if (widget instanceof Notebook) {
-    editor = widget.activeCell?.editor;
+  const { content } = widget;
+
+  if (content instanceof FileEditor) {
+    editor = content.editor;
+  } else if (content instanceof Notebook) {
+    editor = content.activeCell?.editor;
+  }
+
+  if (!(editor instanceof CodeMirrorEditor)) {
+    return undefined;
   }
 
   return editor;
